+++ web/changelog.md
@@ -20,6 +20,7 @@ and start a new "Upcoming" section.
 
 * Port to 1.19.4
 * Add: Manufactory Halo shows active status in its icon (Wormbo)
+* Change: Key of the King's Law no longer hurts items or most other nonliving entities
 * Fix: Ender Air emission not being mentioned in Pure Daisy entry (Wormbo)